Refactor User and Chat model imports: update import paths for User and Chat models to reflect new directory structure, ensuring consistency across authentication and chat functionalities.
