+++ backend/routes/user/__tests__/unit/getAllUsers.test.js
@@ -1,10 +1,10 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import express from 'express';
 import request from 'supertest';
-import User from '../../../../models/User.js';
+import User from '../../../../models/user/User.js';
 
 // Mock the User model
-vi.mock('../../../../models/User.js', () => ({
+vi.mock('../../../../models/user/User.js', () => ({
   default: {
     getAll: vi.fn()
   }
