+++ backend/routes/user/__tests__/unit/resetPassword.test.js
@@ -1,11 +1,11 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import express from 'express';
 import request from 'supertest';
-import User from '../../../../models/User.js';
+import User from '../../../../models/user/User.js';
 import EmailService from '../../../../services/emailService.js';
 
 // Mock dependencies
-vi.mock('../../../../models/User.js', () => ({
+vi.mock('../../../../models/user/User.js', () => ({
   default: {
     findByEmail: vi.fn(),
     storeResetToken: vi.fn(),
