+++ backend/routes/user/get-all-users/controller.js
@@ -1,4 +1,4 @@
-import User from '../../../models/User.js';
+import User from '../../../models/user/User.js';
 
 export const getAllUsers = async (req, res) => {
   try {
