+++ backend/routes/user/update-user/controller.js
@@ -1,4 +1,4 @@
-import User from "../../../models/User.js";
+import User from "../../../models/user/User.js";
 import bcrypt from "bcrypt";
 
 /**
